Implement U+EE01 vertically-centered colon for Android
Android's default system font, Roboto, does not have a calt alternative
to vertically center the colon when it's next to a number. Instead, it
uses a Private Use Area codepoint, U+EE01, for its vertically-centered
colon symbol that is used on the lock screen clock.

This adds U+EE01 as simple glyph that embeds U+E092 (colon.case) as a
component, since that character is a vertically-centered colon --
exactly what we're looking for in this case.

This font is fully usable as an Android system font after this change.